+++ apps/www/src/registry/default/components/editor/plugins/markdown-plugin.tsx
@@ -1,8 +1,14 @@
-import { MarkdownPlugin, remarkMention } from '@udecode/plate-markdown';
+'use client';
+
+import {
+  MarkdownPlugin,
+  remarkMdx,
+  remarkMention,
+} from '@udecode/plate-markdown';
 import { SuggestionPlugin } from '@udecode/plate-suggestion/react';
 import remarkGfm from 'remark-gfm';
 import remarkMath from 'remark-math';
-import remarkMdx from 'remark-mdx';
+
 export const markdownPlugin = MarkdownPlugin.configure({
   options: {
     disallowedNodes: [SuggestionPlugin.key],