+++ packages/core/src/react/components/index.ts
@@ -8,6 +8,7 @@ export * from './EditorMethodsEffect';
 export * from './EditorRefEffect';
 export * from './EditorStateEffect';
 export * from './Plate';
+export * from './PlateContainer';
 export * from './PlateContent';
 export * from './PlateControllerEffect';
 export * from './PlateSlate';