+++ templates/plate-playground-template/src/components/plate-ui/import-toolbar-button.tsx
@@ -44,13 +44,25 @@ export function ImportToolbarButton({ children, ...props }: DropdownMenuProps) {
     return nodes;
   };
 
-  const { openFilePicker } = useFilePicker({
-    accept,
+  const { openFilePicker: openMdFilePicker } = useFilePicker({
+    accept: ['.md'],
     multiple: false,
     onFilesSelected: async ({ plainFiles }) => {
       const text = await plainFiles[0].text();
 
-      const nodes = getFileNodes(text, type);
+      const nodes = getFileNodes(text, 'markdown');
+
+      editor.tf.insertNodes(nodes);
+    },
+  });
+
+  const { openFilePicker: openHtmlFilePicker } = useFilePicker({
+    accept: ['text/html'],
+    multiple: false,
+    onFilesSelected: async ({ plainFiles }) => {
+      const text = await plainFiles[0].text();
+
+      const nodes = getFileNodes(text, 'html');
 
       editor.tf.insertNodes(nodes);
     },
@@ -69,7 +81,7 @@ export function ImportToolbarButton({ children, ...props }: DropdownMenuProps) {
           <DropdownMenuItem
             onSelect={() => {
               setType('html');
-              openFilePicker();
+              openHtmlFilePicker();
             }}
           >
             Import from HTML
@@ -78,7 +90,7 @@ export function ImportToolbarButton({ children, ...props }: DropdownMenuProps) {
           <DropdownMenuItem
             onSelect={() => {
               setType('markdown');
-              openFilePicker();
+              openMdFilePicker();
             }}
           >
             Import from Markdown