+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @udecode/plate-ai
 
+## 47.2.0
+
+### Patch Changes
+
+- [#4236](https://github.com/udecode/plate/pull/4236) by [@felixfeng33](https://github.com/felixfeng33) – use `withoutMdx` option for filtering out.
+
+## 47.1.2
+
+## 47.1.1
+
+## 47.1.0
+
+### Minor Changes
+
+- [#4213](https://github.com/udecode/plate/pull/4213) by [@felixfeng33](https://github.com/felixfeng33) –
+
+  - New `streamInsertChunk` to solve [#4040](https://github.com/udecode/plate/issues/4040) issue where streaming multiple paragraphs was not possible
+  - New `aiChat.rf.removeAnchorAIChat` to remove the AI anchor element after streaming is complete without redo/undo issues
+
+## 47.0.8
+
+## 47.0.7
+
 ## 47.0.6
 
 ## 47.0.5